requirements.txt
flask
flask-cors
numpy
Pillow
gunicorn
torch
torchvision